+++ app/books/add/page.tsx
@@ -1,3 +1,8 @@
+'use client';
+
+import { useRouter } from 'next/navigation';
+import { useSetAtom } from 'jotai';
+import { keywordListAtom } from '@/store/atoms';
 import HeaderLayout from '@/layout/header';
 import DeleteKeywordsButton from '@/ui/books/delete-keywords-button';
 import KeywordList from '@/ui/books/keyword-list';
@@ -18,8 +23,26 @@ export default function AddBook({
     target?: string;
   };
 }) {
+  const router = useRouter();
+  const setKeywordList = useSetAtom(keywordListAtom);
+
   const query = searchParams?.query || '';
   const target = searchParams?.target || '';
+
+  const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
+    event.preventDefault();
+
+    const params = { query, target };
+    const formattedParams = new URLSearchParams(params).toString();
+
+    router.push(`/books/add/list?${formattedParams}`);
+    setKeywordList((prev) => {
+      if (!prev.includes(query)) {
+        return [...prev, query];
+      }
+      return prev;
+    });
+  };
 
   return (
     <>
@@ -27,7 +50,7 @@ export default function AddBook({
 
       {/* Note: `pt-12` - header height만큼 공간 확보 + page section의 자체 패딩 */}
       <section className="pt-16 px-3 pb-4">
-        <SearchForm query={query} target={target} />
+        <SearchForm onSubmit={handleSubmit} />
         <Tab tabs={tabs} />
         <KeywordList />
       </section>
